requirements.txt
bleach==3.1.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
docopt==0.6.2
docutils==0.16
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pkginfo==1.5.0.1
plaid-python==2.4.1
plaid2qif==1.3.2
Pygments==2.6.1
python-dateutil==2.8.0
readme-renderer==25.0
requests==2.23.0
requests-toolbelt==0.9.1
six==1.14.0
tqdm==4.43.0
twine==1.13.0
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.0
